Move WysiwygComposerViewModelProtocol to dedicated file and extract documentation there
