+++ platforms/ios/lib/WysiwygComposer/Sources/WysiwygComposer/Components/WysiwygComposerView/WysiwygComposerViewModelProtocol.swift
@@ -1,0 +1,50 @@
+//
+// Copyright 2022 The Matrix.org Foundation C.I.C
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+import UIKit
+
+public protocol WysiwygComposerViewModelProtocol: AnyObject {
+    /// The textView with placeholder support that the model manages
+    var textView: PlaceholdableTextView? { get set }
+    /// The composer content.
+    var content: WysiwygComposerContent { get }
+
+    /// Update the composer compressed required height if it has changed.
+    ///
+    /// - Parameters:
+    ///   - textView: The composer's text view.
+    func updateCompressedHeightIfNeeded(_ textView: UITextView)
+
+    /// Replace text in the model.
+    ///
+    /// - Parameters:
+    ///   - textView: TextView which currently holds the displayed text in the composer.
+    ///   - range: Range to replace.
+    ///   - replacementText: Replacement text to apply.
+    func replaceText(_ textView: UITextView, range: NSRange, replacementText: String) -> Bool
+
+    /// Select given range of text within the model.
+    ///
+    /// - Parameters:
+    ///   - text: Text currently displayed in the composer.
+    ///   - range: Range to select.
+    func select(text: NSAttributedString, range: NSRange)
+
+    /// Notify that the text view content has changed.
+    ///
+    /// - Parameter textView: The composer's text view.
+    func didUpdateText(textView: UITextView)
+}
